Additional methods for creating Sets.
I have encountered a few cases where I want to create a Set from a slice of
the wrong type. Writing the boilerplate to use slice.Map and then create the
set is just as long (and twice as expensive) as a native loop. Hopefully
having dedicated functions will make it a little more attractive to write
these transformations in functional style.